Capture error messages when an email fails to be sent.
This switches to primarily using the new "wp_mail_succeeded" action. Unfortunately, this action and the "wp_mail_failed" action both pass along a processed list of headers. This prevents us from accessing the Content-Type header, for example, since PHPMail consumes the value. To accommodate this, we still capture the email message using the "wp_mail" filter to retrieve the "raw" unprocessed headers.