Add support for resending an email.
A new Send\Sender interface has been defined to encapsulate the sending of an Email. Optionally, an email can be sent to a new recipient.

A new header action has been added to the email detail screen. When clicked, it opens a form to allow specifying the new "to" address. When submitted, a POST request is made to the new /emails/{uuid}/send API.